Prevent redundant Pydantic model parsing
Exceptions raised by `create_event` are no longer caught within
`create_event` itself. Rather, the caller is expected to handle a
potential `DBAPIError`. In the case of the `/create` endpoint handler,
the error is logged and a 500 response is sent.

This also improves the auto-generated API documentation by enabling it
to be populated with the Pydantic model details.