Add a middleware to require the `X-Fideslog-Version` header
Requiring this header will slightly improve the API request burden, by
short-circuiting any requests made without it. This may include
requests made by bots and/or malicious users. To further improve the
request burden, a companion change will be made at the infrastructure
level, to enforce this header using a WAF.

The `/docs` and `/openapi.json` endpoints are excluded from the
requirement in order to ensure that the auto-generated documentation
remains publicly available. This required header is also intentionally
undocumented, as a mild security buff.