Enable validation of events representing local servers
When submitting events to the `/events` endpoint that represent API
requests made by a server running locally (Ex. a host of `0.0.0.0`),
the endpoint validation fails due to a bug in the third-party
validation library.

These are legitimate events (or if they are illegitimate events, it is
not due to this host value), so they should pass validation. The
validator now simply replaces the `0.0.0.0` host with `localhost`. This
does not affect event accuracy, because the endpoint URL is ultimately
truncated such that it only includes the URL path.